Changed back to one text field.
Changed intent to call the Messaging app explicitly. This solution accomplishes the requirement of recipient choosing but may not be compatible with older versions. The Messaging app must also be set as the default SMS app prior to use.

Problem: Pressing back button on messaging app saves the draft and the containing photo, refusing new photos until the old one is removed from the draft.